Move TraceAction to lib types
